MOve some classees into appropriate packages.
