+++ rdbms/src/main/java/org/spine3/server/storage/jdbc/aggregate/query/AggregateStorageQueryFactory.java
@@ -25,8 +25,8 @@ import org.spine3.server.aggregate.Aggregate;
 import org.spine3.server.aggregate.AggregateEventRecord;
 import org.spine3.server.aggregate.AggregateStorage;
 import org.spine3.server.entity.Visibility;
-import org.spine3.server.storage.jdbc.entity.query.InsertAndMarkEntityQuery;
-import org.spine3.server.storage.jdbc.entity.query.MarkEntityQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.query.InsertAndMarkEntityQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.query.MarkEntityQuery;
 import org.spine3.server.storage.jdbc.entity.visibility.VisibilityHandlingStorageQueryFactory;
 import org.spine3.server.storage.jdbc.entity.visibility.VisibilityQueryFactories;
 import org.spine3.server.storage.jdbc.entity.visibility.query.CreateVisibilityTableQuery;
