+++ rdbms/src/main/java/org/spine3/server/storage/jdbc/entity/query/RecordStorageQueryFactory.java
@@ -29,7 +29,9 @@ import org.spine3.server.storage.RecordStorage;
 import org.spine3.server.storage.jdbc.entity.visibility.VisibilityHandlingStorageQueryFactory;
 import org.spine3.server.storage.jdbc.entity.visibility.VisibilityQueryFactories;
 import org.spine3.server.storage.jdbc.entity.visibility.query.CreateVisibilityTableQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.query.InsertAndMarkEntityQuery;
 import org.spine3.server.storage.jdbc.entity.visibility.query.InsertVisibilityQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.query.MarkEntityQuery;
 import org.spine3.server.storage.jdbc.entity.visibility.query.SelectVisibilityQuery;
 import org.spine3.server.storage.jdbc.entity.visibility.query.UpdateVisibilityQuery;
 import org.spine3.server.storage.jdbc.query.DeleteRowQuery;
