+++ rdbms/src/main/java/org/spine3/server/storage/jdbc/entity/visibility/VisibilityHandler.java
@@ -23,7 +23,7 @@ package org.spine3.server.storage.jdbc.entity.visibility;
 import com.google.common.base.Optional;
 import com.google.protobuf.Message;
 import org.spine3.server.entity.Visibility;
-import org.spine3.server.storage.jdbc.entity.query.MarkEntityQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.query.MarkEntityQuery;
 import org.spine3.server.storage.jdbc.entity.visibility.query.SelectVisibilityQuery;
 
 import static com.google.common.base.Preconditions.checkNotNull;
