+++ rdbms/src/main/java/org/spine3/server/storage/jdbc/entity/visibility/table/VisibilityTable.java
@@ -23,6 +23,8 @@ package org.spine3.server.storage.jdbc.entity.visibility.table;
 import org.spine3.server.entity.Visibility;
 
 /**
+ * A container for the constants adjacent to the table storing the {@linkplain Visibility}.
+ *
  * @author Dmytro Dashenkov.
  */
 public class VisibilityTable {
@@ -33,6 +35,7 @@ public class VisibilityTable {
     public static final int COLUMN_COUNT = 3;
 
     public static final String ID_COL = "id";
+
     public static final int ID_COL_INDEX = 1;
 
     public static final int ARCHIVED_COL_INDEX = 2;
